fix: rename test -> test_utils
