+++ src/crypto/polynomials.rs
@@ -604,7 +604,7 @@ pub fn batch_invert<C: Ciphersuite>(values: &[Scalar<C>]) -> Result<Vec<Scalar<C
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::test::{generate_participants, generate_participants_with_random_ids};
+    use crate::test_utils::{generate_participants, generate_participants_with_random_ids};
     use frost_core::Field;
     use frost_secp256k1::{Secp256K1Group, Secp256K1ScalarField, Secp256K1Sha256};
     use k256::Scalar;
