+++ src/ecdsa/ot_based_ecdsa/test.rs
@@ -6,9 +6,9 @@ use super::{
 };
 use crate::participants::Participant;
 use crate::protocol::Protocol;
-use crate::test::{
+use crate::test_utils::{
     assert_public_key_invariant, generate_participants, generate_participants_with_random_ids,
-    one_coordinator_output, run_keygen, run_protocol, run_refresh, run_reshare,
+    one_coordinator_output, run_keygen, run_protocol, run_refresh, run_reshare, run_sign,
 };
 use crate::{
     crypto::hash::test::scalar_hash_secp256k1, ecdsa::ot_based_ecdsa::RerandomizedPresignOutput,
@@ -18,7 +18,7 @@ use crate::{
         Element, ParticipantList, RerandomizationArguments, Secp256K1Sha256, Signature,
         SignatureOption, Tweak,
     },
-    test::{GenOutput, GenProtocol},
+    test_utils::{GenOutput, GenProtocol},
 };
 
 use rand::rngs::OsRng;
@@ -49,7 +49,7 @@ pub fn run_sign_without_rerandomization(
     let coordinator = participants_presign[index].0;
 
     // run sign instanciation with the necessary arguments
-    let result = crate::test::run_sign::<Secp256K1Sha256, _, _, _>(
+    let result = run_sign::<Secp256K1Sha256, _, _, _>(
         rerand_participants_presign,
         coordinator,
         public_key,
@@ -111,7 +111,7 @@ pub fn run_sign_with_rerandomization(
     let coordinator = participants_presign[index].0;
 
     // run sign instanciation with the necessary arguments
-    let result = crate::test::run_sign::<Secp256K1Sha256, _, _, _>(
+    let result = crate::test_utils::run_sign::<Secp256K1Sha256, _, _, _>(
         rerand_participants_presign,
         coordinator,
         derived_pk,
