+++ src/ecdsa/ot_based_ecdsa/triples/batch_random_ot.rs
@@ -372,7 +372,7 @@ mod test {
     use crate::ecdsa::ot_based_ecdsa::triples::test::run_batch_random_ot;
     use crate::participants::Participant;
     use crate::protocol::internal::{make_protocol, Comms};
-    use crate::test::run_two_party_protocol;
+    use crate::test_utils::run_two_party_protocol;
     use rand_core::OsRng;
 
     #[test]
