+++ src/ecdsa/ot_based_ecdsa/triples/mta.rs
@@ -143,7 +143,7 @@ mod test {
     use crate::errors::ProtocolError;
     use crate::participants::Participant;
     use crate::protocol::internal::{make_protocol, Comms};
-    use crate::test::run_two_party_protocol;
+    use crate::test_utils::run_two_party_protocol;
 
     /// Run the multiplicative to additive protocol
     fn run_mta(
