+++ src/ecdsa/ot_based_ecdsa/triples/multiplication.rs
@@ -227,7 +227,7 @@ mod test {
         ecdsa::ot_based_ecdsa::triples::multiplication::{multiplication, multiplication_many},
         participants::ParticipantList,
         protocol::internal::{make_protocol, Comms},
-        test::{generate_participants, run_protocol, GenProtocol},
+        test_utils::{generate_participants, run_protocol, GenProtocol},
     };
 
     #[test]
