+++ src/ecdsa/robust_ecdsa/sign.rs
@@ -178,7 +178,7 @@ mod test {
         robust_ecdsa::PresignOutput,
         Field, Polynomial, ProjectivePoint, Secp256K1ScalarField,
     };
-    use crate::test::generate_participants;
+    use crate::test_utils::generate_participants;
 
     type PresigSimulationOutput = (Scalar, Polynomial, Polynomial, Polynomial, ProjectivePoint);
 
@@ -301,7 +301,7 @@ mod test {
     #[test]
     fn test_sign_fails_if_s_is_zero() {
         use crate::ecdsa::{ProjectivePoint, Secp256K1ScalarField};
-        use crate::test::generate_participants;
+        use crate::test_utils::generate_participants;
 
         let participants = generate_participants(2);
