+++ src/protocol/echo_broadcast.rs
@@ -329,7 +329,7 @@ mod test {
     use super::*;
     use crate::protocol::internal::{make_protocol, Comms};
     use crate::protocol::{Protocol, ProtocolError};
-    use crate::test::{generate_participants, run_protocol};
+    use crate::test_utils::{generate_participants, run_protocol};
 
     /// This function is similar to `do_broadcast` except it is tailored to
     /// consume the inputs instead of borrowing and become suitable for `make_protocol`
